requirements.txt
alabaster==0.7.12
Babel==2.10.3
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
docutils==0.18.1
idna==3.3
imagesize==1.4.1
imgviz==1.5.0
importlib-metadata==4.12.0
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.2.2
natsort==8.1.0
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
Pygments==2.12.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
QtPy==2.1.0
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
termcolor==1.1.0
urllib3==1.26.10
wincertstore==0.2
zipp==3.8.1
